with editing term scene and updates in code
