+++ src/xcode/MRNBike/Resources/Base.lproj/Localizable.strings
@@ -57,8 +57,9 @@
 "dialogActionGotIt"="Got it!";
 
 //Term Scene
-"termConditionHeader"="Generic Terms and Conditions of MRN Bike mobile application";
-"termConditionText"="Terms and Conditions (\"Terms\") Last updated:  (add date) Please read these Terms and Conditions (\"Terms\", \"Terms and Conditions\") carefully before using the ) websiteandtheM yMobileApp(change this) mobile application (the \"Service\") operated by  My Company (change this)  (\"us\", \"we\", or \"our\"). Your access to and use of the Service is conditioned on your acceptance of and compliance with these Terms. These Terms apply to all visitors, users and others who access or use the Service. By accessing or using the Service you agree to be bound by these Terms. If you disagree with any part of the terms then you may not access the Service. Links To Other Web Sites Our Service may contain links to third-party web sites or services that are not owned or controlled by  My Company (change this). My Company (change this)  has no control over, and assumes no responsibility for, the content, privacy policies, or practices of any third party web sites or services. You further acknowledge and agree that  My Company (change this)  shall not be responsible or liable, directly or indirectly, for any damage or loss caused or alleged to be caused by or in connection with use of or reliance on any such content, goods or services available on or through any such web sites or services. Changes We reserve the right, at our sole discretion, to modify or replace these Terms at any time. If a revision is material we will try to provide at least 3  0 (change this)  days' notice prior to any new terms taking effect. What constitutes a material change will be determined at our sole discretion. Contact Us If you have any questions about these Terms, please contact us.";
+"termConditionHeader"="Terms & Conditions";
+"termConditionText"="Terms and Conditions (\"Terms\") \nLast updated:  (add date) Please read these Terms and Conditions (\"Terms\", \"Terms and Conditions\") carefully before using the ) websiteandtheM yMobileApp(change this) mobile application (the \"Service\") operated by  My Company (change this) Terms and Conditions (\"Terms\") \nLast updated:  (add date) Please read these Terms and Conditions (\"Terms\", \"Terms and Conditions\") carefully before using the ) websiteandtheM yMobileApp(change this) mobil";
+
 "agreeButton"="I AGREE";
 "disagreeButton"="I DISAGREE";
